corrects dependencies
Calsipro had the dependency "syn-bokeh-helpers", which is a private util package. The necessary code is present in calsipro without the dependency, but trying to install calsipro fails without the package available.

This commit removes the wrong dependency and modifies the import statements in calsipro/visualisations to actually access the integrated code.